sandbox_api/test/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-kusto-data==2.0.0
pyodbc
azure-cli-core
#regex
azure-mgmt-kusto==0.6.0
pandas
pytz
azure-storage-blob
natsort
azure.iot.device
requests
azure.eventhub
flashtext